+++ editor/src/external-types.ts
@@ -14,3 +14,8 @@ export type StepEditorProvider = (step: Step, context: StepEditorContext) => HTM
 
 export type StepValidator = (step: Step, _: unknown, definition: Definition) => boolean;
 export type RootValidator = (definition: Definition) => boolean;
+
+export interface ToolboxGroup {
+	name: string;
+	steps: Step[];
+}