+++ editor/src/value-editors/string/string-value-editor.ts
@@ -2,8 +2,8 @@ import { StringValueModel, ValueModelContext } from 'sequential-workflow-editor-
 import { ValueEditor } from '../value-editor';
 import { validationErrorComponent } from '../../components/validation-error-component';
 import { valueEditorContainerComponent } from '../../components/value-editor-container-component';
-import { Html } from '../../core/html';
 import { rowComponent } from '../../components/row-component';
+import { inputComponent } from '../../components/input-component';
 
 export const stringValueEditorId = 'string';
 
@@ -12,18 +12,14 @@ export function stringValueEditor(context: ValueModelContext<StringValueModel>):
 		validation.setDefaultError(context.validate());
 	}
 
-	const input = Html.element('input', {
-		class: 'swe-input swe-stretched',
-		type: 'text'
-	});
-	input.value = context.getValue();
-
-	input.addEventListener('input', () => {
-		context.setValue(input.value);
+	const startValue = context.getValue();
+	const input = inputComponent(startValue);
+	input.onChanged.subscribe(value => {
+		context.setValue(value);
 		validate();
 	});
 
-	const row = rowComponent([input]);
+	const row = rowComponent([input.view]);
 
 	const validation = validationErrorComponent();
 	const container = valueEditorContainerComponent([row.view, validation.view]);